requirements.txt
transformers==4.44.2
datasets==2.21.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
accelerate>=0.30.0
evaluate>=0.4.0
